poc/requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
langchain==0.1.16
langchain-google-vertexai==0.1.3
google-auth==2.29.0
google-api-core==2.18.0
requests==2.31.0
langchain-community==0.0.34
